Move shared selection config components to SelectionConfig & round numbers in LabelledInput
