+++ client/component/HorizontalAxisSelectionConfig.tsx
@@ -1,13 +1,7 @@
 import { LabelledInput } from './LabelledInput';
-import { getSelectionLabel, SELECTION_ICONS } from './selections/utils';
-import { isNumber, toFixedNumber } from './utils';
+import { isNumber } from './utils';
 import HorizontalAxisSelection from './selections/HorizontalAxisSelection';
-import {
-  AlphaInput,
-  ColourPicker,
-  NameInput,
-  XInput,
-} from './SelectionConfigComponents';
+import { XInput } from './SelectionConfigComponents';
 
 interface HorizontalAxisSelectionConfigProps {
   selection: HorizontalAxisSelection;
@@ -19,23 +13,6 @@ export function HorizontalAxisSelectionConfig(
 ) {
   return (
     <>
-      <h4 key="Selection">
-        {' '}
-        {getSelectionLabel(props.selection, SELECTION_ICONS)}{' '}
-      </h4>
-      <ColourPicker
-        selection={props.selection}
-        updateSelections={props.updateSelections}
-      />
-      <NameInput
-        selection={props.selection}
-        updateSelections={props.updateSelections}
-      />
-
-      <AlphaInput
-        selection={props.selection}
-        updateSelections={props.updateSelections}
-      />
       <XInput
         selection={props.selection}
         updateSelections={props.updateSelections}
@@ -43,11 +20,12 @@ export function HorizontalAxisSelectionConfig(
       <LabelledInput<number>
         key="x length"
         label="x length"
-        input={toFixedNumber(props.selection.dimensionLength[0], 5)}
+        input={props.selection.dimensionLength[0]}
         updateValue={(l: number) => {
           props.selection.dimensionLength[0] = l;
           props.updateSelections(props.selection);
         }}
+        decimalPlaces={5}
         isValid={(v) => isNumber(v)}
       />
     </>
