+++ client/component/LinearSelectionConfig.tsx
@@ -1,13 +1,5 @@
-import { getSelectionLabel, SELECTION_ICONS } from './selections/utils';
 import LinearSelection from './selections/LinearSelection';
-import {
-  AlphaInput,
-  AngleInput,
-  ColourPicker,
-  NameInput,
-  XInput,
-  YInput,
-} from './SelectionConfigComponents';
+import { AngleInput, XInput, YInput } from './SelectionConfigComponents';
 
 interface LinearSelectionConfigProps {
   selection: LinearSelection;
@@ -17,26 +9,6 @@ interface LinearSelectionConfigProps {
 export function LinearSelectionConfig(props: LinearSelectionConfigProps) {
   return (
     <>
-      <h4 key="Selection">
-        {' '}
-        {getSelectionLabel(props.selection, SELECTION_ICONS)}{' '}
-      </h4>
-
-      <ColourPicker
-        selection={props.selection}
-        updateSelections={props.updateSelections}
-      />
-
-      <NameInput
-        selection={props.selection}
-        updateSelections={props.updateSelections}
-      />
-
-      <AlphaInput
-        selection={props.selection}
-        updateSelections={props.updateSelections}
-      />
-
       <XInput
         selection={props.selection}
         updateSelections={props.updateSelections}
